chore: Refactor Uno.Themes solution and sample apps for improved runtime testing
