+++ Uno.Themes.sln
@@ -1,3 +1,4 @@
+﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 18
 VisualStudioVersion = 18.2.11415.280
@@ -76,6 +77,20 @@ EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SimpleSampleApp", "src\samples\SimpleSampleApp\SimpleSampleApp.csproj", "{D4E5F6A7-B8C9-0123-DEFA-234567890123}"
 EndProject
 Project("{D954291E-2A0B-460D-934E-DC6B0785DB48}") = "SamplesApp.Shared", "src\samples\SamplesApp.Shared\SamplesApp.Shared.shproj", "{53B0C90E-5C6A-4A6E-AC56-3FCD1CEF5D3F}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "src", "src", "{827E0CD3-B72D-47B6-A68D-7590B98EB39B}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "samples", "samples", "{D1B4CF99-7947-DD6C-81D0-5B5B3989814C}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Uno.Themes.Samples.RuntimeTests", "src\samples\Uno.Themes.Samples.RuntimeTests\Uno.Themes.Samples.RuntimeTests.csproj", "{CBD0978C-3FD0-4296-B65C-CCFC79A5AEED}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "library", "library", "{6C3B6E0B-582A-0347-1D22-231F8AE97553}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Uno.Themes", "Uno.Themes", "{C7A3A149-373E-5088-E8D6-FBC06F8345C9}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Uno.Material", "Uno.Material", "{4AD5D793-3F92-82B9-1A75-5DF0F9E0A70B}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Uno.Cupertino", "Uno.Cupertino", "{B344C419-5E82-B345-73CE-4C64DEF62C78}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -555,6 +570,58 @@ Global
 		{D4E5F6A7-B8C9-0123-DEFA-234567890123}.Release|x64.ActiveCfg = Release|Any CPU
 		{D4E5F6A7-B8C9-0123-DEFA-234567890123}.Release|x64.Build.0 = Release|Any CPU
 		{D4E5F6A7-B8C9-0123-DEFA-234567890123}.Release|x86.ActiveCfg = Release|Any CPU
+		{CBD0978C-3FD0-4296-B65C-CCFC79A5AEED}.Debug|Android.ActiveCfg = Debug|Any CPU
+		{CBD0978C-3FD0-4296-B65C-CCFC79A5AEED}.Debug|Android.Build.0 = Debug|Any CPU
+		{CBD0978C-3FD0-4296-B65C-CCFC79A5AEED}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{CBD0978C-3FD0-4296-B65C-CCFC79A5AEED}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{CBD0978C-3FD0-4296-B65C-CCFC79A5AEED}.Debug|arm64.ActiveCfg = Debug|Any CPU
+		{CBD0978C-3FD0-4296-B65C-CCFC79A5AEED}.Debug|arm64.Build.0 = Debug|Any CPU
+		{CBD0978C-3FD0-4296-B65C-CCFC79A5AEED}.Debug|iPhone.ActiveCfg = Debug|Any CPU
+		{CBD0978C-3FD0-4296-B65C-CCFC79A5AEED}.Debug|iPhone.Build.0 = Debug|Any CPU
+		{CBD0978C-3FD0-4296-B65C-CCFC79A5AEED}.Debug|iPhoneSimulator.ActiveCfg = Debug|Any CPU
+		{CBD0978C-3FD0-4296-B65C-CCFC79A5AEED}.Debug|iPhoneSimulator.Build.0 = Debug|Any CPU
+		{CBD0978C-3FD0-4296-B65C-CCFC79A5AEED}.Debug|macOS.ActiveCfg = Debug|Any CPU
+		{CBD0978C-3FD0-4296-B65C-CCFC79A5AEED}.Debug|macOS.Build.0 = Debug|Any CPU
+		{CBD0978C-3FD0-4296-B65C-CCFC79A5AEED}.Debug|Packages_UWP.ActiveCfg = Debug|Any CPU
+		{CBD0978C-3FD0-4296-B65C-CCFC79A5AEED}.Debug|Packages_UWP.Build.0 = Debug|Any CPU
+		{CBD0978C-3FD0-4296-B65C-CCFC79A5AEED}.Debug|Packages_WinUI.ActiveCfg = Debug|Any CPU
+		{CBD0978C-3FD0-4296-B65C-CCFC79A5AEED}.Debug|Packages_WinUI.Build.0 = Debug|Any CPU
+		{CBD0978C-3FD0-4296-B65C-CCFC79A5AEED}.Debug|Tests.ActiveCfg = Debug|Any CPU
+		{CBD0978C-3FD0-4296-B65C-CCFC79A5AEED}.Debug|Tests.Build.0 = Debug|Any CPU
+		{CBD0978C-3FD0-4296-B65C-CCFC79A5AEED}.Debug|UWP_x64.ActiveCfg = Debug|Any CPU
+		{CBD0978C-3FD0-4296-B65C-CCFC79A5AEED}.Debug|UWP_x64.Build.0 = Debug|Any CPU
+		{CBD0978C-3FD0-4296-B65C-CCFC79A5AEED}.Debug|WASM.ActiveCfg = Debug|Any CPU
+		{CBD0978C-3FD0-4296-B65C-CCFC79A5AEED}.Debug|WASM.Build.0 = Debug|Any CPU
+		{CBD0978C-3FD0-4296-B65C-CCFC79A5AEED}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{CBD0978C-3FD0-4296-B65C-CCFC79A5AEED}.Debug|x64.Build.0 = Debug|Any CPU
+		{CBD0978C-3FD0-4296-B65C-CCFC79A5AEED}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{CBD0978C-3FD0-4296-B65C-CCFC79A5AEED}.Debug|x86.Build.0 = Debug|Any CPU
+		{CBD0978C-3FD0-4296-B65C-CCFC79A5AEED}.Release|Android.ActiveCfg = Release|Any CPU
+		{CBD0978C-3FD0-4296-B65C-CCFC79A5AEED}.Release|Android.Build.0 = Release|Any CPU
+		{CBD0978C-3FD0-4296-B65C-CCFC79A5AEED}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{CBD0978C-3FD0-4296-B65C-CCFC79A5AEED}.Release|Any CPU.Build.0 = Release|Any CPU
+		{CBD0978C-3FD0-4296-B65C-CCFC79A5AEED}.Release|arm64.ActiveCfg = Release|Any CPU
+		{CBD0978C-3FD0-4296-B65C-CCFC79A5AEED}.Release|arm64.Build.0 = Release|Any CPU
+		{CBD0978C-3FD0-4296-B65C-CCFC79A5AEED}.Release|iPhone.ActiveCfg = Release|Any CPU
+		{CBD0978C-3FD0-4296-B65C-CCFC79A5AEED}.Release|iPhone.Build.0 = Release|Any CPU
+		{CBD0978C-3FD0-4296-B65C-CCFC79A5AEED}.Release|iPhoneSimulator.ActiveCfg = Release|Any CPU
+		{CBD0978C-3FD0-4296-B65C-CCFC79A5AEED}.Release|iPhoneSimulator.Build.0 = Release|Any CPU
+		{CBD0978C-3FD0-4296-B65C-CCFC79A5AEED}.Release|macOS.ActiveCfg = Release|Any CPU
+		{CBD0978C-3FD0-4296-B65C-CCFC79A5AEED}.Release|macOS.Build.0 = Release|Any CPU
+		{CBD0978C-3FD0-4296-B65C-CCFC79A5AEED}.Release|Packages_UWP.ActiveCfg = Release|Any CPU
+		{CBD0978C-3FD0-4296-B65C-CCFC79A5AEED}.Release|Packages_UWP.Build.0 = Release|Any CPU
+		{CBD0978C-3FD0-4296-B65C-CCFC79A5AEED}.Release|Packages_WinUI.ActiveCfg = Release|Any CPU
+		{CBD0978C-3FD0-4296-B65C-CCFC79A5AEED}.Release|Packages_WinUI.Build.0 = Release|Any CPU
+		{CBD0978C-3FD0-4296-B65C-CCFC79A5AEED}.Release|Tests.ActiveCfg = Release|Any CPU
+		{CBD0978C-3FD0-4296-B65C-CCFC79A5AEED}.Release|Tests.Build.0 = Release|Any CPU
+		{CBD0978C-3FD0-4296-B65C-CCFC79A5AEED}.Release|UWP_x64.ActiveCfg = Release|Any CPU
+		{CBD0978C-3FD0-4296-B65C-CCFC79A5AEED}.Release|UWP_x64.Build.0 = Release|Any CPU
+		{CBD0978C-3FD0-4296-B65C-CCFC79A5AEED}.Release|WASM.ActiveCfg = Release|Any CPU
+		{CBD0978C-3FD0-4296-B65C-CCFC79A5AEED}.Release|WASM.Build.0 = Release|Any CPU
+		{CBD0978C-3FD0-4296-B65C-CCFC79A5AEED}.Release|x64.ActiveCfg = Release|Any CPU
+		{CBD0978C-3FD0-4296-B65C-CCFC79A5AEED}.Release|x64.Build.0 = Release|Any CPU
+		{CBD0978C-3FD0-4296-B65C-CCFC79A5AEED}.Release|x86.ActiveCfg = Release|Any CPU
+		{CBD0978C-3FD0-4296-B65C-CCFC79A5AEED}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -577,6 +644,12 @@ Global
 		{C3D4E5F6-A7B8-9012-CDEF-123456789012} = {F42C65EA-6E6A-43B9-9D9F-D8C6063AD309}
 		{D4E5F6A7-B8C9-0123-DEFA-234567890123} = {F42C65EA-6E6A-43B9-9D9F-D8C6063AD309}
 		{53B0C90E-5C6A-4A6E-AC56-3FCD1CEF5D3F} = {F42C65EA-6E6A-43B9-9D9F-D8C6063AD309}
+		{D1B4CF99-7947-DD6C-81D0-5B5B3989814C} = {827E0CD3-B72D-47B6-A68D-7590B98EB39B}
+		{CBD0978C-3FD0-4296-B65C-CCFC79A5AEED} = {D1B4CF99-7947-DD6C-81D0-5B5B3989814C}
+		{6C3B6E0B-582A-0347-1D22-231F8AE97553} = {827E0CD3-B72D-47B6-A68D-7590B98EB39B}
+		{C7A3A149-373E-5088-E8D6-FBC06F8345C9} = {6C3B6E0B-582A-0347-1D22-231F8AE97553}
+		{4AD5D793-3F92-82B9-1A75-5DF0F9E0A70B} = {6C3B6E0B-582A-0347-1D22-231F8AE97553}
+		{B344C419-5E82-B345-73CE-4C64DEF62C78} = {6C3B6E0B-582A-0347-1D22-231F8AE97553}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {66F0EE99-4E63-43F2-9504-0CFC104E7084}
