+++ src/samples/MaterialSampleApp/MaterialSampleApp.csproj
@@ -37,6 +37,10 @@
     <UnoFeatures>
       SkiaRenderer;
     </UnoFeatures>
+
+    <!-- Uno.UI.RuntimeTests.Engine (pulled in transitively via Uno.Themes.Samples.RuntimeTests)
+         still references the legacy DOM-based WebAssembly runtime; bypass the SDK check. -->
+    <UnoDisableValidateNoUnoWasmDOMRuntimeDependency>true</UnoDisableValidateNoUnoWasmDOMRuntimeDependency>
   </PropertyGroup>
 
 
@@ -44,19 +48,13 @@
 
   <ItemGroup>
     <ProjectReference Include="../../library/Uno.Material/Uno.Material.WinUI.csproj" />
+    <!-- Shared runtime-tests library: hosts all [TestClass] definitions so they run in every sample. -->
+    <ProjectReference Include="../Uno.Themes.Samples.RuntimeTests/Uno.Themes.Samples.RuntimeTests.csproj" />
   </ItemGroup>
 
   <ItemGroup>
     <PackageReference Include="Uno.ShowMeTheXAML" />
     <PackageReference Include="Uno.ShowMeTheXAML.MSBuild" />
-    <PackageReference Include="Uno.UI.RuntimeTests.Engine" />
-    <!--
-      Required for hot-reload runtime tests ([RunsInSecondaryApp]) — enables HAS_UNO_DEVSERVER
-      in Uno.UI.RuntimeTests.Engine. Debug-only: in Release the DevServer assembly is not
-      published and would otherwise cause a FileNotFoundException at app startup.
-    -->
-    <PackageReference Include="Uno.WinUI.DevServer" PrivateAssets="all" Condition="'$(Configuration)' == 'Debug'" />
-    <PackageReference Include="MSTest.TestFramework" />
   </ItemGroup>
 
   <ItemGroup>
