+++ src/samples/Uno.Themes.Samples.RuntimeTests/GlobalUsings.cs
@@ -1,0 +1,11 @@
+global using System;
+global using System.Collections.Generic;
+global using System.Linq;
+global using System.Threading;
+global using System.Threading.Tasks;
+global using Microsoft.UI.Xaml;
+global using Microsoft.UI.Xaml.Controls;
+global using Microsoft.UI.Xaml.Media;
+global using Microsoft.VisualStudio.TestTools.UnitTesting;
+global using Uno.UI.RuntimeTests;
+global using Windows.UI;
